Moved Grok helper code to a separate Gradle module and let ingest-common module depend on it.
