+++ libs/grok-common/src/main/resources/patterns/aws
