+++ libs/grok-common/src/main/resources/patterns/grok-patterns
