+++ libs/grok-common/src/main/resources/patterns/linux-syslog
