+++ libs/grok-common/src/main/resources/patterns/mcollective-patterns
